Remove all fortran global spectral analysis routines from pyshtools. Divide shclasses into several files (not working yet...)
