+++ examples/python/TimingAccuracy/TimingAccuracyGLQC.py
@@ -12,7 +12,7 @@ import numpy as np
 
 sys.path.append(os.path.join(os.path.dirname(__file__), "../../.."))
 from pyshtools import expand
-from pyshtools import shtools
+from pyshtools import spectralanalysis
 
 
 # ==== MAIN FUNCTION ====
@@ -46,7 +46,7 @@ def TimingAccuracyGLQC():
     cilm.real = np.random.normal(loc=0., scale=1.,
                                  size=(2, maxdeg + 1, maxdeg + 1))
 
-    old_power = shtools.SHPowerSpectrumC(cilm)
+    old_power = spectralanalysis.spectrum(cilm)
     new_power = 1. / (1. + ls)**beta  # initialize degrees > 0 to power-law
     cilm[:, :, :] *= np.sqrt(new_power / old_power)[None, :, None]
     cilm[~mask] = 0.
