+++ pyshtools/shclasses/__init__.py
@@ -1,0 +1,45 @@
+"""
+Class interface for SHTOOLS.
+
+pyshtools defines several classes that facilitate the interactive
+examination of geographical gridded data and their associated
+spherical harmonic coefficients. Subclasses are used to handle different
+internal data types and superclasses are used to implement interface
+functions and documentation.
+
+pyshtools class structure:
+
+    SHCoeffs
+        SHRealCoeffs
+        SHComplexCoeffs
+
+    SHGrid
+        DHRealGrid
+        DHComplexGrid
+        GLQRealGrid
+        GLQComplexGrid
+
+    SHWindow
+        SHWindowCap
+        SHWindowMask
+
+For more information, see the documentation for the top level classes.
+"""
+
+from __future__ import absolute_import as _absolute_import
+from __future__ import division as _division
+from __future__ import print_function as _print_function
+
+from .shcoeffs import SHCoeffs
+from .shcoeffs import SHRealCoeffs
+from .shcoeffs import SHComplexCoeffs
+
+from .shgrid import SHGrid
+from .shgrid import DHRealGrid
+from .shgrid import DHComplexGrid
+from .shgrid import GLQRealGrid
+from .shgrid import GLQComplexGrid
+
+from .shwindow import SHWindow
+from .shwindow import SHWindowCap
+from .shwindow import SHWindowMask
