+++ pyshtools/shclasses/shgrid.py
@@ -1,0 +1,936 @@
+"""
+    Spherical Harmonic Grid classes
+
+        SHGrid: DHRealGrid, DHComplexGrid, GLQRealGrid, GLQComplexGrid
+"""
+from __future__ import absolute_import as _absolute_import
+from __future__ import division as _division
+from __future__ import print_function as _print_function
+
+import numpy as _np
+import matplotlib as _mpl
+import matplotlib.pyplot as _plt
+import copy as _copy
+
+from .. import shtools as _shtools
+from ..spectralanalysis import spectrum as _spectrum
+
+from .shcoeffs import *
+
+
+__all__ = ['SHGrid', 'DHRealGrid', 'DHComplexGrid', 'GLQRealGrid',
+           'GLQComplexGrid']
+
+
+class SHGrid(object):
+    """
+    Class for spatial gridded data on the sphere.
+
+    Grids can be initialized from:
+
+    >>> x = SHGrid.from_array(array)
+    >>> x = SHGrid.from_file('fname.dat')
+
+    The class instance defines the following class attributes:
+
+    data       : Gridded array of the data.
+    nlat, nlon : The number of latitude and longitude bands in the grid.
+    lmax       : The maximum spherical harmonic degree that can be resolved
+                 by the grid sampling.
+    sampling   : For Driscoll and Healy grids, the longitudinal sampling
+                 of the grid. Either nlong=nlat or nlong=2*nlat.
+    kind       : Either 'complex' or 'real' for the data type.
+    grid       : Either 'DH' or 'GLQ' for Driscoll and Healy grids or Gauss-
+                 Legendre Quadrature grids.
+    zeros      : The cos(colatitude) nodes used with Gauss-Legendre
+                 Quadrature grids. Default is None.
+    weights    : The latitudinal weights used with Gauss-Legendre
+                 Quadrature grids. Default is None.
+
+    Each class instance provides the following methods:
+
+    to_array()   : Return the raw gridded data as a numpy array.
+    to_file()   : Save gridded data to a text or binary file.
+    lats()      : Return a vector containing the latitudes of each row
+                  of the gridded data.
+    lons()      : Return a vector containing the longitudes of each column
+                  of the gridded data.
+    expand()    : Expand the grid into spherical harmonics.
+    copy()      : Return a copy of the class instance.
+    plot()      : Plot the raw data using a simple cylindrical projection.
+    plot3d      : Plot the raw data on a 3d sphere.
+    info()      : Print a summary of the data stored in the SHGrid instance.
+    """
+
+    def __init__():
+        """Unused constructor of the super class."""
+        print('Initialize the class using one of the class methods:\n'
+              '>>> SHGrid.from_array?\n'
+              '>>> SHGrid.from_file?\n')
+
+    # ---- factory methods
+    @classmethod
+    def from_array(self, array, grid='DH', copy=True):
+        """
+        Initialize the class instance from an input array.
+
+        Usage
+        -----
+        x = SHGrid.from_array(array, [grid, copy])
+
+        Returns
+        -------
+        x : SHGrid class instance
+
+        Parameters
+        ----------
+        array : ndarray, shape (nlat, nlon)
+            2-D numpy array of the gridded data, where nlat and nlon are the
+            number of latitudinal and longitudinal bands, respectively.
+        grid : str, optional, default = 'DH'
+            'DH' or 'GLQ' for Driscoll and Healy grids or Gauss Legendre
+            Quadrature grids, respectively.
+        copy : bool, optional, default = True
+            If True (default), make a copy of array when initializing the class
+            instance. If False, initialize the class instance with a reference
+            to array.
+        """
+        if _np.iscomplexobj(array):
+            kind = 'complex'
+        else:
+            kind = 'real'
+
+        if type(grid) != str:
+            raise ValueError('grid must be a string. ' +
+                             'Input type was {:s}'
+                             .format(str(type(grid))))
+
+        if grid.upper() not in set(['DH', 'GLQ']):
+            raise ValueError(
+                "grid must be 'DH' or 'GLQ'. Input value was {:s}."
+                .format(repr(grid))
+                )
+
+        for cls in self.__subclasses__():
+            if cls.istype(kind) and cls.isgrid(grid):
+                return cls(array, copy=copy)
+
+    @classmethod
+    def from_file(self, fname, binary=False, **kwargs):
+        """
+        Initialize the class instance from gridded data in a file.
+
+        Usage
+        -----
+        x = SHGrid.from_file(fname, [binary, **kwargs])
+
+        Returns
+        -------
+        x : SHGrid class instance
+
+        Parameters
+        ----------
+        fname : str
+            The filename containing the gridded data. For text files (default)
+            the file is read using the numpy routine loadtxt(), whereas for
+            binary files, the file is read using numpy.load(). The dimensions
+            of the array must be nlon=nlat or nlon=2*nlat for Driscoll and
+            Healy grids, or nlon=2*nlat-1 for Gauss-Legendre Quadrature grids.
+        binary : bool, optional, default = False
+            If False, read a text file. If True, read a binary 'npy' file.
+        **kwargs : keyword arguments, optional
+            Keyword arguments of numpy.loadtxt() or numpy.load().
+        """
+        if binary is False:
+            data = _np.loadtxt(fname, **kwargs)
+        elif binary is True:
+            data = _np.load(fname, **kwargs)
+        else:
+            raise ValueError('binary must be True or False. '
+                             'Input value is {:s}'.format(binary))
+
+        if _np.iscomplexobj(data):
+            kind = 'complex'
+        else:
+            kind = 'real'
+
+        if (data.shape[1] == data.shape[0]) or (data.shape[1] ==
+                                                2 * data.shape[0]):
+            grid = 'DH'
+        elif data.shape[1] == 2 * data.shape[0] - 1:
+            grid = 'GLQ'
+        else:
+            raise ValueError('Input grid must be dimensioned as ' +
+                             '(nlat, nlon). For DH grids, nlon = nlat or ' +
+                             'nlon = 2 * nlat. For GLQ grids, nlon = ' +
+                             '2 * nlat - 1. Input dimensions are nlat = ' +
+                             '{:d}, nlon = {:d}'.format(data.shape[0],
+                                                        data.shape[1]))
+
+        for cls in self.__subclasses__():
+            if cls.istype(kind) and cls.isgrid(grid):
+                return cls(data)
+
+    def copy(self):
+        """Return a deep copy of the class instance."""
+        return _copy.deepcopy(self)
+
+    def to_file(self, filename, binary=False, **kwargs):
+        """
+        Save gridded data to a file.
+
+        Usage
+        -----
+        x.to_file(filename, [binary, **kwargs])
+
+        Parameters
+        ----------
+        filename : str
+            Name of output file. For text files (default), the file will be
+            saved automatically in gzip compressed format if the filename ends
+            in .gz.
+        binary : bool, optional, default = False
+            If False, save as text using numpy.savetxt(). If True, save as a
+            'npy' binary file using numpy.save().
+        **kwargs : keyword arguments, optional
+            Keyword arguments of numpy.savetxt() and numpy.save().
+        """
+        if binary is False:
+            _np.savetxt(filename, self.data, **kwargs)
+        elif binary is True:
+            _np.save(filename, self.data, **kwargs)
+        else:
+            raise ValueError('binary must be True or False. '
+                             'Input value is {:s}'.format(binary))
+
+    # ---- operators ----
+    def __add__(self, other):
+        """Add two similar grids or a grid and a scaler: self + other."""
+        if isinstance(other, SHGrid):
+            if (self.grid == other.grid and self.data.shape ==
+                    other.data.shape and self.kind == other.kind):
+                data = self.data + other.data
+                return SHGrid.from_array(data, grid=self.grid)
+            else:
+                raise ValueError('The two grids must be of the ' +
+                                 'same kind and have the same shape.')
+        elif _np.isscalar(other) is True:
+            data = self.data + other
+            return SHGrid.from_array(data, grid=self.grid)
+        else:
+            raise NotImplementedError('Mathematical operator not implemented' +
+                                      'for these operands.')
+
+    def __radd__(self, other):
+        """Add two similar grids or a grid and a scaler: self + other."""
+        return self.__add__(other)
+
+    def __sub__(self, other):
+        """Subtract two similar grids or a grid and a scaler: self - other."""
+        if isinstance(other, SHGrid):
+            if (self.grid == other.grid and self.data.shape ==
+                    other.data.shape and self.kind == other.kind):
+                data = self.data - other.data
+                return SHGrid.from_array(data, grid=self.grid)
+            else:
+                raise ValueError('The two grids must be of the ' +
+                                 'same kind and have the same shape.')
+        elif _np.isscalar(other) is True:
+            data = self.data - other
+            return SHGrid.from_array(data, grid=self.grid)
+        else:
+            raise NotImplementedError('Mathematical operator not implemented' +
+                                      'for these operands.')
+
+    def __rsub__(self, other):
+        """Subtract two similar grids or a grid and a scaler: other - self."""
+        if isinstance(other, SHGrid):
+            if (self.grid == other.grid and self.data.shape ==
+                    other.data.shape and self.kind == other.kind):
+                data = other.data - self.data
+                return SHGrid.from_array(data, grid=self.grid)
+            else:
+                raise ValueError('The two grids must be of the ' +
+                                 'same kind and have the same shape.')
+        elif _np.isscalar(other) is True:
+            data = other - self.data
+            return SHGrid.from_array(data, grid=self.grid)
+        else:
+            raise NotImplementedError('Mathematical operator not implemented' +
+                                      'for these operands.')
+
+    def __mul__(self, other):
+        """Multiply two similar grids or a grid and a scaler: self * other."""
+        if isinstance(other, SHGrid):
+            if (self.grid == other.grid and self.data.shape ==
+                    other.data.shape and self.kind == other.kind):
+                data = self.data * other.data
+                return SHGrid.from_array(data, grid=self.grid)
+            else:
+                raise ValueError('The two grids must be of the ' +
+                                 'same kind and have the same shape.')
+        elif _np.isscalar(other) is True:
+            data = self.data * other
+            return SHGrid.from_array(data, grid=self.grid)
+        else:
+            raise NotImplementedError('Mathematical operator not implemented' +
+                                      'for these operands.')
+
+    def __rmul__(self, other):
+        """Multiply two similar grids or a grid and a scaler: other * self."""
+        return self.__mul__(other)
+
+    def __div__(self, other):
+        """
+        Divide two similar grids or a grid and a scalar, when
+        __future__.division is not in effect.
+        """
+        if isinstance(other, SHGrid):
+            if (self.grid == other.grid and self.data.shape ==
+                    other.data.shape and self.kind == other.kind):
+                data = self.data / other.data
+                return SHGrid.from_array(data, grid=self.grid)
+            else:
+                raise ValueError('The two grids must be of the ' +
+                                 'same kind and have the same shape.')
+        elif _np.isscalar(other) is True:
+            data = self.data / other
+            return SHGrid.from_array(data, grid=self.grid)
+        else:
+            raise NotImplementedError('Mathematical operator not implemented' +
+                                      'for these operands.')
+
+    def __truediv__(self, other):
+        """
+        Divide two similar grids or a grid and a scalar, when
+        __future__.division is in effect.
+        """
+        if isinstance(other, SHGrid):
+            if (self.grid == other.grid and self.data.shape ==
+                    other.data.shape and self.kind == other.kind):
+                data = self.data / other.data
+                return SHGrid.from_array(data, grid=self.grid)
+            else:
+                raise ValueError('The two grids must be of the ' +
+                                 'same kind and have the same shape.')
+        elif _np.isscalar(other) is True:
+            data = self.data / other
+            return SHGrid.from_array(data, grid=self.grid)
+        else:
+            raise NotImplementedError('Mathematical operator not implemented' +
+                                      'for these operands.')
+
+    def __pow__(self, other):
+        """Raise a grid to a scalar power: pow(self, other)."""
+        if _np.isscalar(other) is True:
+            data = pow(self.data, other)
+            return SHGrid.from_array(data, grid=self.grid)
+        else:
+            raise NotImplementedError('Mathematical operator not implemented' +
+                                      'for these operands.')
+
+    # ---- Extract grid properties ----
+    def lats(self, degrees=True):
+        """
+        Return the latitudes of each row of the gridded data.
+
+        Usage
+        -----
+        lats = x.lats([degrees])
+
+        Returns
+        -------
+        lats : ndarray, shape (nlat)
+            1-D numpy array of size nlat containing the latitude of each row
+            of the gridded data.
+
+        Parameters
+        -------
+        degrees : bool, optional, default = True
+            If True, the output will be in degrees. If False, the output will
+            be in radians.
+        """
+        if degrees is False:
+            return _np.radians(self._lats())
+        else:
+            return self._lats()
+
+    def lons(self, degrees=True):
+        """
+        Return the longitudes of each column of the gridded data.
+
+        Usage
+        -----
+        lons = x.get_lon([degrees])
+
+        Returns
+        -------
+        lons : ndarray, shape (nlon)
+            1-D numpy array of size nlon containing the longitude of each row
+            of the gridded data.
+
+        Parameters
+        -------
+        degrees : bool, optional, default = True
+            If True, the output will be in degrees. If False, the output will
+            be in radians.
+        """
+        if degrees is False:
+            return _np.radians(self._lons())
+        else:
+            return self._lons()
+
+    def to_array(self):
+        """
+        Return the raw gridded data as a numpy array.
+
+        Usage
+        -----
+        grid = x.to_array()
+
+        Returns
+        -------
+        grid : ndarray, shape (nlat, nlon)
+            2-D numpy array of the gridded data.
+        """
+        return _np.copy(self.data)
+
+    def plot3d(self, show=True, fname=None, elevation=0, azimuth=0):
+        """
+        Plot the raw data on a 3d sphere.
+
+        This routines becomes slow for large grids because it is based on
+        matplotlib3d.
+
+        Usage
+        -----
+        x.plot3d([show, fname])
+
+        Parameters
+        ----------
+        show : bool, optional, default = True
+            If True, plot the image to the screen.
+        fname : str, optional, default = None
+            If present, save the image to the file.
+        """
+        from mpl_toolkits.mplot3d import Axes3D  # NOQA
+
+        nlat, nlon = self.nlat, self.nlon
+        cmap = _plt.get_cmap('RdBu_r')
+
+        if self.kind == 'real':
+            data = self.data
+        elif self.kind == 'complex':
+            data = _np.abs(self.data)
+        else:
+            raise ValueError('Grid has to be either real or complex, not {}'
+                             .format(self.kind))
+
+        lats = self.lats()
+        lons = self.lons()
+
+        if self.grid == 'DH':
+            # add south pole
+            lats_circular = _np.append(lats, [-90.])
+        elif self.grid == 'GLQ':
+            # add north and south pole
+            lats_circular = _np.hstack(([90.], lats, [-90.]))
+        lons_circular = _np.append(lons, [lons[0]])
+
+        nlats_circular = len(lats_circular)
+        nlons_circular = len(lons_circular)
+
+        sshape = nlats_circular, nlons_circular
+
+        # make uv sphere and store all points
+        u = _np.radians(lons_circular)
+        v = _np.radians(90. - lats_circular)
+
+        x = _np.sin(v)[:, None] * _np.cos(u)[None, :]
+        y = _np.sin(v)[:, None] * _np.sin(u)[None, :]
+        z = _np.cos(v)[:, None] * _np.ones_like(lons_circular)[None, :]
+
+        points = _np.vstack((x.flatten(), y.flatten(), z.flatten()))
+
+        # fill data for all points. 0 lon has to be repeated (circular mesh)
+        # and the south pole has to be added in the DH grid
+        if self.grid == 'DH':
+            magn_point = _np.zeros((nlat + 1, nlon + 1))
+            magn_point[:-1, :-1] = data
+            magn_point[-1, :] = _np.mean(data[-1])  # not exact !
+            magn_point[:-1, -1] = data[:, 0]
+        if self.grid == 'GLQ':
+            magn_point = _np.zeros((nlat + 2, nlon + 1))
+            magn_point[1:-1, :-1] = data
+            magn_point[0, :] = _np.mean(data[0])  # not exact !
+            magn_point[-1, :] = _np.mean(data[-1])  # not exact !
+            magn_point[1:-1, -1] = data[:, 0]
+
+        # compute face color, which is the average of all neighbour points
+        magn_face = 1./4. * (magn_point[1:, 1:] + magn_point[:-1, 1:] +
+                             magn_point[1:, :-1] + magn_point[:-1, :-1])
+
+        magnmax_face = _np.max(_np.abs(magn_face))
+        magnmax_point = _np.max(_np.abs(magn_point))
+
+        # compute colours and displace the points
+        norm = _plt.Normalize(-magnmax_face / 2., magnmax_face / 2., clip=True)
+        colors = cmap(norm(magn_face.flatten()))
+        colors = colors.reshape(nlats_circular - 1, nlons_circular - 1, 4)
+        points *= (1. + magn_point.flatten() / magnmax_point / 2.)
+        x = points[0].reshape(sshape)
+        y = points[1].reshape(sshape)
+        z = points[2].reshape(sshape)
+
+        # plot 3d radiation pattern
+        fig = _plt.figure(figsize=(10, 10))
+        ax3d = fig.add_subplot(1, 1, 1, projection='3d')
+        ax3d.plot_surface(x, y, z, rstride=1, cstride=1, facecolors=colors)
+        ax3d.set(xlim=(-1.5, 1.5), ylim=(-1.5, 1.5), zlim=(-1.5, 1.5),
+                 xticks=[-1, 1], yticks=[-1, 1], zticks=[-1, 1])
+        ax3d.set_axis_off()
+        ax3d.view_init(elev=elevation, azim=azimuth)
+
+        # show or save output
+        if show:
+            _plt.show()
+        if fname is not None:
+            fig.savefig(fname)
+
+        return fig, ax3d
+
+    # ---- Plotting routines ----
+    def plot(self, show=True, fname=None):
+        """
+        Plot the raw data using a simple cylindrical projection.
+
+        Usage
+        -----
+        x.plot([show, fname])
+
+        Parameters
+        ----------
+        show : bool, optional, default = True
+            If True, plot the image to the screen.
+        fname : str, optional, default = None
+            If present, save the image to the file.
+        """
+        fig, ax = self._plot()
+        if show:
+            _plt.show()
+        if fname is not None:
+            fig.savefig(fname)
+        return fig, ax
+
+    def expand(self, normalization='4pi', csphase=1, **kwargs):
+        """
+        Expand the grid into spherical harmonics.
+
+        Usage
+        -----
+        clm = x.expand([normalization, csphase, lmax_calc])
+
+        Returns
+        -------
+        clm : SHCoeffs class instance
+
+        Parameters
+        ----------
+        normalization : str, optional, default = '4pi'
+            Normalization of the spherical harmonic coefficients: '4pi' for
+            geodesy 4-pi normalized, 'ortho' for orthonormalized, or 'schmidt'
+            for Schmidt semi-normalized.
+        csphase : int, optional, default = 1
+            Condon-Shortley phase convention: 1 to exclude the phase factor,
+            or -1 to include it.
+        lmax_calc : int, optional, default = x.lmax
+            Maximum spherical harmonic degree to return.
+        """
+        if type(normalization) != str:
+            raise ValueError('normalization must be a string. ' +
+                             'Input type was {:s}'
+                             .format(str(type(normalization))))
+
+        if normalization.lower() not in set(['4pi', 'ortho', 'schmidt']):
+            raise ValueError(
+                "The normalization must be '4pi', 'ortho' " +
+                "or 'schmidt'. Input value was {:s}."
+                .format(repr(normalization))
+                )
+
+        if csphase != 1 and csphase != -1:
+            raise ValueError(
+                "csphase must be either 1 or -1. Input value was {:s}."
+                .format(repr(csphase))
+                )
+
+        return self._expand(normalization=normalization, csphase=csphase,
+                            **kwargs)
+
+    def info(self):
+        """
+        Print a summary of the data stored in the SHGrid instance.
+
+        Usage
+        -----
+        x.info()
+        """
+        print('kind = {:s}\ngrid = {:s}\n'.format(repr(self.kind),
+                                                  repr(self.grid)), end='')
+        if self.grid == 'DH':
+            print('sampling = {:d}\n'.format(self.sampling), end='')
+        print('nlat = {:d}\nnlon = {:d}\nlmax = {:d}'.format(self.nlat,
+                                                             self.nlon,
+                                                             self.lmax))
+
+
+# ---- Real Driscoll and Healy grid class ----
+
+class DHRealGrid(SHGrid):
+    """Class for real Driscoll and Healy (1994) grids."""
+
+    @staticmethod
+    def istype(kind):
+        return kind == 'real'
+
+    @staticmethod
+    def isgrid(grid):
+        return grid == 'DH'
+
+    def __init__(self, array, copy=True):
+        self.nlat, self.nlon = array.shape
+
+        if self.nlat % 2 != 0:
+            raise ValueError('Input arrays for DH grids must have an even ' +
+                             'number of latitudes: nlat = {:d}'
+                             .format(self.nlat)
+                             )
+        if self.nlon == 2 * self.nlat:
+            self.sampling = 2
+        elif self.nlat == self.nlon:
+            self.sampling = 1
+        else:
+            raise ValueError('Input array has shape (nlat={:d},nlon={:d})\n'
+                             .format(self.nlat, self.nlon) +
+                             'but needs nlat=nlon or nlat=2*nlon'
+                             )
+
+        self.lmax = int(self.nlat / 2 - 1)
+        self.grid = 'DH'
+        self.kind = 'real'
+
+        if copy:
+            self.data = _np.copy(array)
+        else:
+            self.data = array
+
+    def _lats(self):
+        """Return the latitudes (in degrees) of the gridded data."""
+        lats = _np.linspace(90.0, -90.0 + 180.0 / self.nlat, num=self.nlat)
+        return lats
+
+    def _lons(self):
+        """Return the longitudes (in degrees) of the gridded data."""
+        lons = _np.linspace(0.0, 360.0 - 360.0 / self.nlon, num=self.nlon)
+        return lons
+
+    def _expand(self, normalization, csphase, **kwargs):
+        """Expand the grid into real spherical harmonics."""
+        if normalization.lower() == '4pi':
+            norm = 1
+        elif normalization.lower() == 'schmidt':
+            norm = 2
+        elif normalization.lower() == 'ortho':
+            norm = 4
+        else:
+            raise ValueError(
+                "The normalization must be '4pi', 'ortho' " +
+                "or 'schmidt'. Input value was {:s}."
+                .format(repr(normalization))
+                )
+
+        cilm = _shtools.SHExpandDH(self.data, norm=norm, csphase=csphase,
+                                   sampling=self.sampling,
+                                   **kwargs)
+        coeffs = SHCoeffs.from_array(cilm,
+                                     normalization=normalization.lower(),
+                                     csphase=csphase, copy=False)
+        return coeffs
+
+    def _plot(self):
+        """Plot the raw data using a simply cylindrical projection."""
+        fig, ax = _plt.subplots(1, 1)
+        ax.imshow(self.data, origin='upper', extent=(0., 360., -90., 90.))
+        ax.set_xlabel('longitude')
+        ax.set_ylabel('latitude')
+        fig.tight_layout(pad=0.5)
+        return fig, ax
+
+
+# ---- Complex Driscoll and Healy grid class ----
+class DHComplexGrid(SHGrid):
+    """
+    Class for complex Driscoll and Healy (1994) grids.
+    """
+    @staticmethod
+    def istype(kind):
+        return kind == 'complex'
+
+    @staticmethod
+    def isgrid(grid):
+        return grid == 'DH'
+
+    def __init__(self, array, copy=True):
+        self.nlat, self.nlon = array.shape
+
+        if self.nlat % 2 != 0:
+            raise ValueError('Input arrays for DH grids must have an even ' +
+                             'number of latitudes: nlat = {:d}'
+                             .format(self.nlat)
+                             )
+        if self.nlon == 2 * self.nlat:
+            self.sampling = 2
+        elif self.nlat == self.nlon:
+            self.sampling = 1
+        else:
+            raise ValueError('Input array has shape (nlat={:d},nlon={:d})\n' +
+                             'but needs nlat=nlon or nlat=2*nlon'
+                             .format(self.nlat, self.nlon)
+                             )
+
+        self.lmax = int(self.nlat / 2 - 1)
+        self.grid = 'DH'
+        self.kind = 'complex'
+
+        if copy:
+            self.data = _np.copy(array)
+        else:
+            self.data = array
+
+    def _lats(self):
+        """
+        Return a vector containing the latitudes (in degrees) of each row
+        of the gridded data.
+        """
+        lats = _np.linspace(90.0, -90.0 + 180.0 / self.nlat, num=self.nlat)
+        return lats
+
+    def _lons(self):
+        """
+        Return a vector containing the longitudes (in degrees) of each row
+        of the gridded data.
+        """
+        lons = _np.linspace(0., 360.0 - 360.0 / self.nlon, num=self.nlon)
+        return lons
+
+    def _expand(self, normalization, csphase, **kwargs):
+        """Expand the grid into real spherical harmonics."""
+        if normalization.lower() == '4pi':
+            norm = 1
+        elif normalization.lower() == 'schmidt':
+            norm = 2
+        elif normalization.lower() == 'ortho':
+            norm = 4
+        else:
+            raise ValueError(
+                "The normalization must be '4pi', 'ortho' " +
+                "or 'schmidt'. Input value was {:s}."
+                .format(repr(normalization))
+                )
+
+        cilm = _shtools.SHExpandDHC(self.data, norm=norm, csphase=csphase,
+                                    **kwargs)
+        coeffs = SHCoeffs.from_array(cilm,
+                                     normalization=normalization.lower(),
+                                     csphase=csphase, copy=False)
+        return coeffs
+
+    def _plot(self):
+        """Plot the raw data using a simply cylindrical projection."""
+        fig, ax = _plt.subplots(2, 1)
+        ax.flat[0].imshow(self.data.real, origin='upper',
+                          extent=(0., 360., -90., 90.))
+        ax.flat[0].set_title('Real component')
+        ax.flat[0].set_xlabel('longitude')
+        ax.flat[0].set_ylabel('latitude')
+        ax.flat[1].imshow(self.data.imag, origin='upper',
+                          extent=(0., 360., -90., 90.))
+        ax.flat[1].set_title('Imaginary component')
+        ax.flat[1].set_xlabel('longitude')
+        ax.flat[1].set_ylabel('latitude')
+        fig.tight_layout(pad=0.5)
+        return fig, ax
+
+
+# ---- Real Gaus Legendre Quadrature grid class ----
+
+class GLQRealGrid(SHGrid):
+    """
+    Class for real Gauss-Legendre Quadrature grids.
+    """
+    @staticmethod
+    def istype(kind):
+        return kind == 'real'
+
+    @staticmethod
+    def isgrid(grid):
+        return grid == 'GLQ'
+
+    def __init__(self, array, zeros=None, weights=None, copy=True):
+        self.nlat, self.nlon = array.shape
+        self.lmax = self.nlat - 1
+
+        if self.nlat != self.lmax + 1 or self.nlon != 2 * self.lmax + 1:
+            raise ValueError('Input array has shape (nlat={:d}, nlon={:d})\n' +
+                             'but needs (nlat={:d}, {:d})'
+                             .format(self.nlat, self.nlon, self.lmax+1,
+                                     2*self.lmax+1)
+                             )
+
+        if zeros is None or weights is None:
+            self.zeros, self.weights = _shtools.SHGLQ(self.lmax)
+        else:
+            self.zeros = zeros
+            self.weights = weights
+
+        self.grid = 'GLQ'
+        self.kind = 'real'
+        if copy:
+            self.data = _np.copy(array)
+        else:
+            self.data = array
+
+    def _lats(self):
+        """
+        Return a vector containing the latitudes (in degrees) of each row
+        of the gridded data.
+        """
+        lats = 90. - _np.arccos(self.zeros) * 180. / _np.pi
+        return lats
+
+    def _lons(self):
+        """
+        Return a vector containing the longitudes (in degrees) of each column
+        of the gridded data.
+        """
+        lons = _np.linspace(0.0, 360.0 - 360.0 / self.nlon, num=self.nlon)
+        return lons
+
+    def _expand(self, normalization, csphase, **kwargs):
+        """Expand the grid into real spherical harmonics."""
+        if normalization.lower() == '4pi':
+            norm = 1
+        elif normalization.lower() == 'schmidt':
+            norm = 2
+        elif normalization.lower() == 'ortho':
+            norm = 4
+        else:
+            raise ValueError(
+                "The normalization must be '4pi', 'ortho' " +
+                "or 'schmidt'. Input value was {:s}."
+                .format(repr(normalization))
+                )
+
+        cilm = _shtools.SHExpandGLQ(self.data, self.weights, self.zeros,
+                                    norm=norm, csphase=csphase, **kwargs)
+        coeffs = SHCoeffs.from_array(cilm,
+                                     normalization=normalization.lower(),
+                                     csphase=csphase,
+                                     copy=False)
+        return coeffs
+
+    def _plot(self):
+        """Plot the raw data using a simply cylindrical projection."""
+
+        fig, ax = _plt.subplots(1, 1)
+        ax.imshow(self.data, origin='upper')
+        ax.set_xlabel('GLQ longitude index')
+        ax.set_ylabel('GLQ latitude index')
+        fig.tight_layout(pad=0.5)
+        return fig, ax
+
+
+# ---- Complex Gaus Legendre Quadrature grid class ----
+
+class GLQComplexGrid(SHGrid):
+    """
+    Class for complex Gauss Legendre Quadrature grids.
+    """
+    @staticmethod
+    def istype(kind):
+        return kind == 'complex'
+
+    @staticmethod
+    def isgrid(grid):
+        return grid == 'GLQ'
+
+    def __init__(self, array, zeros=None, weights=None, copy=True):
+        self.nlat, self.nlon = array.shape
+        self.lmax = self.nlat - 1
+
+        if self.nlat != self.lmax + 1 or self.nlon != 2 * self.lmax + 1:
+            raise ValueError('Input array has shape (nlat={:d}, nlon={:d})\n' +
+                             'but needs (nlat={:d}, {:d})'
+                             .format(self.nlat, self.nlon, self.lmax+1,
+                                     2*self.lmax+1)
+                             )
+
+        if zeros is None or weights is None:
+            self.zeros, self.weights = _shtools.SHGLQ(self.lmax)
+        else:
+            self.zeros = zeros
+            self.weights = weights
+
+        self.grid = 'GLQ'
+        self.kind = 'complex'
+
+        if copy:
+            self.data = _np.copy(array)
+        else:
+            self.data = array
+
+    def _lats(self):
+        """Return the latitudes (in degrees) of the gridded data rows."""
+        lats = 90. - _np.arccos(self.zeros) * 180. / _np.pi
+        return lats
+
+    def _lons(self):
+        """Return the longitudes (in degrees) of the gridded data columns."""
+        lons = _np.linspace(0., 360. - 360. / self.nlon, num=self.nlon)
+        return lons
+
+    def _expand(self, normalization, csphase, **kwargs):
+        """Expand the grid into real spherical harmonics."""
+        if normalization.lower() == '4pi':
+            norm = 1
+        elif normalization.lower() == 'schmidt':
+            norm = 2
+        elif normalization.lower() == 'ortho':
+            norm = 4
+        else:
+            raise ValueError(
+                "The normalization must be '4pi', 'ortho' " +
+                "or 'schmidt'. Input value was {:s}."
+                .format(repr(normalization))
+                )
+
+        cilm = _shtools.SHExpandGLQC(self.data, self.weights, self.zeros,
+                                     norm=norm, csphase=csphase, **kwargs)
+        coeffs = SHCoeffs.from_array(cilm,
+                                     normalization=normalization.lower(),
+                                     csphase=csphase, copy=False)
+        return coeffs
+
+    def _plot(self):
+        """Plot the raw data using a simply cylindrical projection."""
+        fig, ax = _plt.subplots(2, 1)
+        ax.flat[0].imshow(self.data.real, origin='upper')
+        ax.flat[0].set_title('Real component')
+        ax.flat[0].set_xlabel('longitude index')
+        ax.flat[0].set_ylabel('latitude index')
+        ax.flat[1].imshow(self.data.imag, origin='upper')
+        ax.flat[1].set_title('Imaginary component')
+        ax.flat[1].set_xlabel('GLQ longitude index')
+        ax.flat[1].set_ylabel('GLQ latitude index')
+        fig.tight_layout(pad=0.5)
+        return fig, ax
